tests/python/requirements.txt
asyncio==3.4.3
asyncpg==0.29.0
psycopg2-binary==2.9.9
aiopg==1.4.0
